backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
groq==0.4.1
python-dotenv==1.0.0
aiosqlite==0.19.0
alembic==1.13.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
